arrange selectors and actions
