+++ catalogue/webapp/test/e2e/actions/search.ts
@@ -1,21 +1,13 @@
+import { searchImagesForm } from '../selectors/images';
 import {
-  modalexpandedImaged,
-  modalexpandedImageViewMoreButton,
-} from '../selectors/images';
-import {
-  searchImagesForm,
-  searchWorksForm,
-  worksSearchInputField,
-  colourSelectorFilterDropDown,
-  mobileModal,
-  formatFilterMobileButton,
-  colourSelector,
-  mobileModalCloseButton,
-  formatFilterDropDownButton,
   formatFilterDropDown,
-  imagesResultsListItem,
+  formatFilterDropDownButton,
+  formatFilterMobileButton,
+  mobileModal,
+  mobileModalCloseButton,
+  worksSearchInputField,
 } from '../selectors/search';
-
+import { searchWorksForm } from '../selectors/works';
 import { isMobile, fillInputAction, pressEnterAction } from './common';
 
 type formType = 'images' | 'works';
@@ -71,28 +63,4 @@ export async function clickActionCloseModalFilterButton(
     condition === 'images' ? searchImagesForm : searchWorksForm
   } ${mobileModalCloseButton}`;
   await page.click(selector);
-}
-
-// images
-
-export async function clickActionColourDropDown(): Promise<void> {
-  await page.click(colourSelectorFilterDropDown);
-}
-
-export async function clickActionColourPicker(): Promise<void> {
-  await page.click(colourSelector, {
-    position: { x: 100, y: 100 },
-  });
-}
-
-export async function clickActionClickImageResultItem(
-  nthChild: number
-): Promise<void> {
-  await page.click(`${imagesResultsListItem}:nth-child(${nthChild}) a`);
-}
-
-export async function clickActionClickViewExpandedImage(): Promise<void> {
-  await page.click(
-    `${modalexpandedImaged} ${modalexpandedImageViewMoreButton}`
-  );
 }
