+++ catalogue/webapp/test/e2e/selectors/search.ts
@@ -1,21 +1,20 @@
-export const searchImagesForm =
-  'form[aria-describedby="images-form-description"]';
-export const worksSearchInputField = 'input[aria-label="search input field"]';
-export const workSearchResultsContainer = 'ul[role="list"]';
-export const worksSearchResultsListItem = `${workSearchResultsContainer} li[role="listitem"]`;
-export const formatFilterDropDown = '#formats';
-export const formatFilterDropDownButton = 'button[aria-controls="formats"]';
-export const colourSelectorFilterDropDown = `${searchImagesForm} button[aria-controls="images.color"]`;
+// input
 
+export const worksSearchInputField = 'input[aria-label="search input field"]';
+
+// modal
+
+export const mobileModal = '#mobile-filters-modal';
+export const mobileModalCloseButton = `${mobileModal} button[aria-label="close filter button"]`;
 export const formatFilterDropDownContainer = 'div[id="format"]';
 export const formatFilterMobileButton =
   'button[aria-controls="mobile-filters-modal"]';
-export const mobileModal = '#mobile-filters-modal';
-export const mobileModalImageSearch = `${searchImagesForm} ${mobileModal}`;
-export const mobileModalCloseButton = `${mobileModal} button[aria-label="close filter button"]`;
-export const searchWorksForm =
-  'form[aria-describedby="library-catalogue-form-description"]';
-export const colourSelector = '.react-colorful div[aria-label="Color"]';
 
-export const imagesResultsContainer = 'ul[role="list"]';
-export const imagesResultsListItem = `${imagesResultsContainer} li[role="listitem"]`;
+// dropdown
+
+export const formatFilterDropDown = '#formats';
+export const formatFilterDropDownButton = 'button[aria-controls="formats"]';
+
+// result list
+
+export const searchResultsContainer = 'ul[role="list"]';
