--- catalogue/webapp/test/e2e/selectors/work.ts
@@ -1,1 +1,0 @@
-export const workTitleHeading = 'h1[role="heading"]';
